pyproject.toml
[project]
name = "symbolic-recursion"
version = "0.1.0"
description = "Symbolic Memory Core — proto-AGI skeleton with novelty/confidence harness"
requires-python = ">=3.9"
readme = "README.md"
dependencies = ["numpy", "sentence-transformers>=2.2.2"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["symbolic_recursion"]
include-package-data = true

[project.optional-dependencies]
faiss = ["faiss-cpu"]
